Update server.py
Imported render_template in order to render the html upon server startup. There is also a function saveScore I made to try and get something to happen when I clicked the begin button. Comes as a GET as of now however, but it does display the message. Kept Raf's code with the long endpoint just in case.